torrent/torrent_service/requirements.txt
Django==5.1.6
djangorestframework==3.15.2
drf-nested-routers==0.93.5
gunicorn==23.0.0
psycopg==3.2.4
psycopg-binary==3.2.4
requests==2.32.3
sqlparse==0.5.3
django-cors-headers
libtorrent==2.0.11
ffmpeg-python==0.2.0
srt-to-vtt==1.4.0